vehicle_speed/requirements.txt
ultralytics==8.3.32
opencv-contrib-python==4.10.0.84
numpy==2.1.3
PyYAML==6.0.2